Extract segmentation implementation to its own source file
